\(Q_3\) Statistic of Yen (1984) for Testlets

This function calculates the average \(Q_3\) statistic (Yen, 1984) within and between testlets.

Usage

Q3.testlet(q3.res, testlet.matrix, progress=TRUE)

Arguments

q3.res

An object generated by Q3

testlet.matrix

A matrix with two columns. The first column contains names of the testlets and the second names of the items. See the examples for the definition of such matrices.

progress

Logical indicating whether computation progress should be displayed.

Value

A list with following entries

testlet.q3

Data frame with average \(Q_3\) statistics within testlets

testlet.q3.korr

Matrix of average \(Q_3\) statistics within and between testlets

References

Yen, W. M. (1984). Effects of local item dependence on the fit and equating performance of the three-parameter logistic model. Applied Psychological Measurement, 8, 125-145.

See also

Examples

#############################################################################
# EXAMPLE 1: data.read. The 12 items are arranged in 4 testlets
#############################################################################
data(data.read)

# estimate the Rasch model
mod <- sirt::rasch.mml2( data.read)
mod$item

# estmate WLEs
mod.wle <- sirt::wle.rasch( dat=data.read, b=mod$item$b )

# Yen's Q3 statistic
mod.q3 <- sirt::Q3( dat=data.read, theta=mod.wle$theta, b=mod$item$b )

# Yen's Q3 statistic with testlets
items <- colnames(data.read)
testlet.matrix <- cbind( substring(  items,1,1), items )
mod.testletq3 <- sirt::Q3.testlet( q3.res=mod.q3,testlet.matrix=testlet.matrix)
mod.testletq3