Jackknife Estimation of NOHARM Analysis

This function performs a jackknife estimation of NOHARM analysis to get standard errors based on a replication method (see Christoffersson, 1977).

R2noharm.jackknife(object, jackunits=NULL)

# S3 method for R2noharm.jackknife
summary(object, logfile=NULL, ...)

Arguments

object
Object of class R2noharm

jackunits
A vector of integers or a number. If it is a number, then it refers to the number of jackknife units. If it is a vector of integers, then this vector defines the allocation of persons jackknife units. Integers corresponds to row indexes in the data set.

logfile
File name if the summary should be sunk into a file

...
Further arguments to be passed

Value

A list of lists with following entries:

partable

Data frame with parameters

se.pars

List of estimated standard errors for all parameter estimates: tanaka.stat, rmsr.stat, rmsea.stat, chisquare_df.stat, thresholds.stat, final.constants.stat, uniquenesses.stat, factor.cor.stat, loadings.stat, loadings.theta.stat

jackknife.pars

List with obtained results by jackknifing for all parameters: j.tanaka, j.rmsr, rmsea, chisquare_df, j.pm, j.thresholds, j.factor.cor, j.loadings, j.loadings.theta

u.jacknunits

Unique jackknife elements

References

Christoffersson, A. (1977). Two-step weighted least squares factor analysis of dichotomized variables. Psychometrika, 42, 433-438.

See also