Exploratory DETECT Analysis

This function estimates the DETECT index (Stout, Habing, Douglas & Kim, 1996; Zhang & Stout, 1999a, 1999b) in an exploratory way. Conditional covariances of itempairs are transformed into a distance matrix such that items are clustered by the hierarchical Ward algorithm (Roussos, Stout & Marden, 1998). Note that the function will not provide the same output as the original DETECT software.

Usage

expl.detect(data, score, nclusters, N.est=NULL, seed=NULL, bwscale=1.1,
    smooth=TRUE, use_sum_score=FALSE, hclust_method="ward.D", estsample=NULL)

Arguments

data

An \(N \times I\) data frame of dichotomous or polytomous responses. Missing responses are allowed.

score

An ability estimate, e.g. the WLE, sum score or mean score

nclusters

Maximum number of clusters used in the exploratory analysis

N.est

Number of students in a (possible) validation of the DETECT index. N.est students are drawn at random from data.

seed

Random seed

bwscale

Bandwidth scale factor

smooth

Logical indicating whether smoothing should be applied for conditional covariance estimation

use_sum_score

Logical indicating whether sum score should be used. With this option, the bias corrected conditional covariance of Zhang and Stout (1999) is used.

hclust_method

Clustering method used as the argument method in stats::hclust.

estsample

Optional vector of subject indices that defines the estimation sample

Value

A list with following entries

detect.unweighted

Unweighted DETECT statistics

detect.weighted

Weighted DETECT statistics. Weighting is done proportionally to sample sizes of item pairs.

clusterfit

Fit of the cluster method

itemcluster

Cluster allocations

use_sum_score

References

Roussos, L. A., Stout, W. F., & Marden, J. I. (1998). Using new proximity measures with hierarchical cluster analysis to detect multidimensionality. Journal of Educational Measurement, 35, 1-30.

Stout, W., Habing, B., Douglas, J., & Kim, H. R. (1996). Conditional covariance-based nonparametric multidimensionality assessment. Applied Psychological Measurement, 20, 331-354.

Zhang, J., & Stout, W. (1999a). Conditional covariance structure of generalized compensatory multidimensional items, Psychometrika, 64, 129-152.

Zhang, J., & Stout, W. (1999b). The theoretical DETECT index of dimensionality and its application to approximate simple structure, Psychometrika, 64, 213-249.

See also