MCMC Estimation of the Hierarchical IRT Model for Criterion-Referenced Measurement

This function estimates the hierarchical IRT model for criterion-referenced measurement which is based on a two-parameter normal ogive response function (Janssen, Tuerlinckx, Meulders & de Boeck, 2000).

Usage

mcmc.2pnoh(dat, itemgroups, prob.mastery=c(.5,.8), weights=NULL,
      burnin=500, iter=1000, N.sampvalues=1000,
      progress.iter=50, prior.variance=c(1,1), save.theta=FALSE)

Arguments

dat

Data frame with dichotomous item responses

itemgroups

Vector with characters or integers which define the criterion to which an item is associated.

prob.mastery

Probability levels which define nonmastery, transition and mastery stage (see Details)

weights

An optional vector with student sample weights

burnin

Number of burnin iterations

iter

Total number of iterations

N.sampvalues

Maximum number of sampled values to save

progress.iter

Display progress every progress.iter-th iteration. If no progress display is wanted, then choose progress.iter larger than iter.

prior.variance

Scale parameter of the inverse gamma distribution for the \(\sigma^2\) and \(\nu^2\) item variance parameters

save.theta

Should theta values be saved?

Details

The hierarchical IRT model for criterion-referenced measurement (Janssen et al., 2000) assumes that every item \(i\) intends to measure a criterion \(k\). The item response function is defined as $$ P(X_{pik}=1 | \theta_p )= \Phi [ \alpha_{ik} ( \theta_p - \beta_{ik} ) ] \quad, \quad \theta_p \sim N(0,1) $$ Item parameters \((\alpha_{ik},\beta_{ik})\) are hierarchically modeled, i.e. $$ \beta_{ik} \sim N( \xi_k, \sigma^2 ) \quad \mbox{and} \quad \alpha_{ik} \sim N( \omega_k, \nu^2 ) $$ In the mcmc.list output object, also the derived parameters \(d_{ik}=\alpha_{ik} \beta_{ik}\) and \(\tau_k=\xi_k \omega_k\) are calculated. Mastery and nonmastery probabilities are based on a reference item \(Y_{k}\) of criterion \(k\) and a response function $$ P(Y_{pk}=1 | \theta_p )= \Phi [ \omega_{k} ( \theta_p - \xi_{k} ) ] \quad, \quad \theta_p \sim N(0,1) $$ With known item parameters and person parameters, response probabilities of criterion \(k\) are calculated. If a response probability of criterion \(k\) is larger than prob.mastery[2], then a student is defined as a master. If this probability is smaller than prob.mastery[1], then a student is a nonmaster. In all other cases, students are in a transition stage.

In the mcmcobj output object, the parameters d[i] are defined by \(d_{ik}=\alpha_{ik} \cdot \beta_{ik}\) while tau[k] are defined by \( \tau_k=\xi_k \cdot \omega_k \).

Value

A list of class mcmc.sirt with following entries:

mcmcobj

Object of class mcmc.list

summary.mcmcobj

Summary of the mcmcobj object. In this summary the Rhat statistic and the mode estimate MAP is included. The variable PercSEratio indicates the proportion of the Monte Carlo standard error in relation to the total standard deviation of the posterior distribution.

burnin

Number of burnin iterations

iter

Total number of iterations

alpha.chain

Sampled values of \(\alpha_{ik}\) parameters

beta.chain

Sampled values of \(\beta_{ik}\) parameters

xi.chain

Sampled values of \(\xi_{k}\) parameters

omega.chain

Sampled values of \(\omega_{k}\) parameters

sigma.chain

Sampled values of \(\sigma\) parameter

nu.chain

Sampled values of \(\nu\) parameter

theta.chain

Sampled values of \(\theta_p\) parameters

deviance.chain

Sampled values of Deviance values

EAP.rel

EAP reliability

person

Data frame with EAP person parameter estimates for \(\theta_p\) and their corresponding posterior standard deviations

dat

Used data frame

weights

Used student weights

...

Further values

References

Janssen, R., Tuerlinckx, F., Meulders, M., & De Boeck, P. (2000). A hierarchical IRT model for criterion-referenced measurement. Journal of Educational and Behavioral Statistics, 25, 285-306.

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: Simulated data according to Janssen et al. (2000, Table 2)
#############################################################################

N <- 1000
Ik <- c(4,6,8,5,9,6,8,6,5)
xi.k <- c( -.89, -1.13, -1.23, .06, -1.41, -.66, -1.09, .57, -2.44)
omega.k <- c(.98, .91, .76, .74, .71, .80, .79, .82, .54)

# select 4 attributes
K <- 4
Ik <- Ik[1:K] ; xi.k <- xi.k[1:K] ; omega.k <- omega.k[1:K]
sig2 <- 3.02
nu2 <- .09
I <- sum(Ik)
b <- rep( xi.k, Ik ) + stats::rnorm(I, sd=sqrt(sig2) )
a <- rep( omega.k, Ik ) + stats::rnorm(I, sd=sqrt(nu2) )
theta1 <- stats::rnorm(N)
t1 <- rep(1,N)
p1 <- stats::pnorm( outer(t1,a) * ( theta1 - outer(t1,b) ) )
dat <- 1  * ( p1 > stats::runif(N*I)  )
itemgroups <- rep( paste0("A", 1:K ), Ik )

# estimate model
mod <- sirt::mcmc.2pnoh(dat, itemgroups, burnin=200, iter=1000 )
# summary
summary(mod)
# plot
plot(mod$mcmcobj, ask=TRUE)
# write coda files
mcmclist2coda( mod$mcmcobj, name="simul_2pnoh" )
}