Nonparametric Estimation of Item Response Functions

This function does nonparametric item response function estimation (Ramsay, 1991).

np.dich(dat, theta, thetagrid, progress=FALSE, bwscale=1.1,
       method="normal")

Arguments

dat
An \(N \times I\) data frame of dichotomous item responses

theta
Estimated theta values, for example weighted likelihood estimates from wle.rasch

thetagrid
A vector of theta values where the nonparametric item response functions shall be evaluated.

progress
Display progress?

bwscale
The bandwidth parameter \(h\) is calculated by the formula \(h=\)bwscale\(\cdot N^{-1/5}\)

method
The default normal performs kernel regression with untransformed item responses. The method binomial uses nonparametric logistic regression implemented in the sm library.

Value

A list with following entries

dat

Original data frame

thetagrid

Vector of theta values at which the item response functions are evaluated

theta

Used theta values as person parameter estimates

estimate

Estimated item response functions

...

References

Ramsay, J. O. (1991). Kernel smoothing approaches to nonparametric item characteristic curve estimation. Psychometrika, 56, 611-630.

Examples

#############################################################################
# EXAMPLE 1: Reading dataset
#############################################################################
data( data.read )
dat <- data.read

# estimate Rasch model
mod <- sirt::rasch.mml2( dat )
# WLE estimation
wle1 <- sirt::wle.rasch( dat=dat, b=mod$item$b )$theta
# nonparametric function estimation
np1 <- sirt::np.dich( dat=dat, theta=wle1, thetagrid=seq(-2.5, 2.5, len=100 ) )
print( str(np1))
# plot nonparametric item response curves
plot( np1, b=mod$item$b )