Estimation of a NOHARM Analysis from within R

This function enables the estimation of a NOHARM analysis (Fraser & McDonald, 1988; McDonald, 1982a, 1982b, 1997) from within R. NOHARM estimates a compensatory multidimensional factor analysis for dichotomous response data. Arguments of this function strictly follow the rules of the NOHARM manual (see Fraser & McDonald, 2012; Lee & Lee, 2016).

Usage

R2noharm(dat=NULL,pm=NULL, n=NULL, model.type, weights=NULL, dimensions=NULL,
      guesses=NULL, noharm.path, F.pattern=NULL, F.init=NULL,
      P.pattern=NULL, P.init=NULL, digits.pm=4, writename=NULL,
      display.fit=5,  dec=".", display=TRUE)

# S3 method for R2noharm
summary(object, logfile=NULL, ...)

Arguments

dat

An \(N \times I\) data frame of item responses for \(N\) subjects and \(I\) items

pm

A matrix or a vector containing product-moment correlations

n

Sample size. This value must only be included if pm is provided.

model.type

Can be "EFA" (exploratory factor analysis) or "CFA" (confirmatory factor analysis).

weights

Optional vector of student weights

dimensions

Number of dimensions in exploratory factor analysis

guesses

An optional vector of fixed guessing parameters of length \(I\). In case of the default NULL, all guessing parameters are set to zero.

noharm.path

Local path where the NOHARM 4 command line 64-bit version is located.

F.pattern

Pattern matrix for \(F\) (\(I \times D\))

F.init

Initial matrix for \(F\) (\(I \times D\))

P.pattern

Pattern matrix for \(P\) (\(D \times D\))

P.init

Initial matrix for \(P\) (\(D \times D\))

digits.pm

Number of digits after decimal separator which are used for estimation

writename

Name for NOHARM input and output files

display.fit

How many digits (after decimal separator) should be used for printing results on the R console?

dec

Decimal separator ("." or ",")

display

Display output?

object

Object of class R2noharm

logfile

File name if the summary should be sunk into a file

...

Further arguments to be passed

Details

NOHARM estimates a multidimensional compensatory item response model with the probit link function \(\Phi\). For item responses \(X_{pi}\) of person \(p\) on item \(i\) the model equation is defined as $$P( X_{pi}=1 | \bold{\theta}_p )=c_i + ( 1 - c_i ) \Phi( f_{i0} + f_{i1} \theta_{p1} + ... + f_{iD} \theta_{pD} ) $$ where \(F=(f_{id})\) is a loading matrix and \(P\) the covariance matrix of \(\bold{\theta}_p\). The guessing parameters \(c_i\) must be provided as fixed values.

For the definition of \(F\) and \(P\) matrices, please consult the NOHARM manual.

This function needs the 64-bit command line version which can be downloaded from (some links may be broken in the meantime)

http://noharm.niagararesearch.ca/nh4cldl.html
https://noharm.software.informer.com/4.0/
https://cehs.unl.edu/edpsych/software-urls-and-other-interesting-sites/

Value

A list with following entries

tanaka

Tanaka index

rmsr

RMSR statistic

N.itempair

Sample sizes of pairwise item observations

pm

Product moment matrix

weights

Used student weights

guesses

Fixed guessing parameters

residuals

Residual covariance matrix

final.constants

Vector of final constants

thresholds

Threshold parameters

uniquenesses

Item uniquenesses

loadings.theta

Matrix of loadings in theta parametrization (common factor parametrization)

factor.cor

Covariance matrix of factors

difficulties

Item difficulties (for unidimensional models)

discriminations

Item discriminations (for unidimensional models)

loadings

Loading matrix (latent trait parametrization)

model.type

Used model type

Nobs

Number of observations

Nitems

Number of items

modtype

Model type according to the NOHARM specification (see NOHARM manual)

F.init

Initial loading matrix for \(F\)

F.pattern

Pattern loading matrix for \(F\)

P.init

Initial covariance matrix for \(P\)

P.pattern

Pattern covariance matrix for \(P\)

dat

Original data frame

systime

System time

noharm.path

Used NOHARM directory

digits.pm

Number of digits in product moment matrix

dec

Used decimal symbol

display.fit

Number of digits for fit display

dimensions

Number of dimensions

chisquare

Statistic \(\chi^2\)

Nestpars

Number of estimated parameters

df

Degrees of freedom

chisquare_df

Ratio \(\chi^2 / df\)

rmsea

RMSEA statistic

p.chisquare

Significance for \(\chi^2\) statistic

References

Fraser, C., & McDonald, R. P. (1988). NOHARM: Least squares item factor analysis. Multivariate Behavioral Research, 23, 267-269. https://doi.org/10.1207/s15327906mbr2302_9

Fraser, C., & McDonald, R. P. (2012). NOHARM 4 Manual.
http://noharm.niagararesearch.ca/nh4man/nhman.html.

Lee, J. J., & Lee, M. K. (2016). An overview of the normal ogive harmonic analysis robust method (NOHARM) approach to item response theory. Tutorials in Quantitative Methods for Psychology, 12(1), 1-8. https://doi.org/10.20982/tqmp.12.1.p001

McDonald, R. P. (1982a). Linear versus nonlinear models in item response theory. Applied Psychological Measurement, 6(4), 379-396. doi:10.1177/014662168200600402

McDonald, R. P. (1982b). Unidimensional and multidimensional models for item response theory. I.R.T., C.A.T. conference, Minneapolis, 1982, Proceedings.

McDonald, R. P. (1997). Normal-ogive multidimensional model. In W. van der Linden & R. K. Hambleton (1997): Handbook of modern item response theory (pp. 257-269). New York: Springer. http://dx.doi.org/10.1007/978-1-4757-2691-6

Note

Possible errors often occur due to wrong dec specification.

See also

Examples