Extended Bradley-Terry Model

The function btm estimates an extended Bradley-Terry model (Hunter, 2004; see Details). Parameter estimation uses a bias corrected joint maximum likelihood estimation method based on \(\varepsilon\)-adjustment (see Bertoli-Barsotti, Lando & Punzo, 2014). See Details for the algorithm.

The function btm_sim simulated data from the extended Bradley-Terry model.

btm(data, judge=NULL, ignore.ties=FALSE, fix.eta=NULL, fix.delta=NULL, fix.theta=NULL,
       maxiter=100, conv=1e-04, eps=0.3, wgt.ties=.5)

# S3 method for btm
summary(object, file=NULL, digits=4,...)

# S3 method for btm
predict(object, data=NULL, ...)

btm_sim(theta, eta=0, delta=-99, repeated=FALSE)

Arguments

data
Data frame with three columns. The first two columns contain labels from the units in the pair comparison. The third column contains the result of the comparison. "1" means that the first units wins, "0" means that the second unit wins and "0.5" means a draw (a tie).

judge
Optional vector of judge identifiers (if multiple judges are available)

ignore.ties
Logical indicating whether ties should be ignored.

fix.eta
Numeric value for a fixed \(\eta\) value

fix.delta
Numeric value for a fixed \(\delta\) value

fix.theta
A vector with entries for fixed theta values.

maxiter
Maximum number of iterations

conv
Convergence criterion

eps
The \(\varepsilon\) parameter for the \(\varepsilon\)-adjustment method (see Bertoli-Barsotti, Lando & Punzo, 2014) which reduces bias in ability estimates. In case of \(\varepsilon=0\), persons with extreme scores are removed from the pairwise comparison.

wgt.ties
Weighting parameter for ties, see formula in Details. The default is .5

object
Object of class btm

file
Optional file name for sinking the summary into

digits
Number of digits after decimal to print

...
Further arguments to be passed.

theta
Vector of abilities

eta
Value of \(\eta\) parameter

delta
Value of \(\delta\) parameter

repeated
Logical indicating whether repeated ratings of dyads (for home advantage effect) should be simulated

Details

The extended Bradley-Terry model for the comparison of individuals \(i\) and \(j\) is defined as $$P(X_{ij}=1 ) \propto \exp( \eta + \theta_i ) $$ $$P(X_{ij}=0 ) \propto \exp( \theta_j ) $$ $$P(X_{ij}=0.5) \propto \exp( \delta + w_T ( \eta + \theta_i +\theta_j ) ) $$ The parameters \(\theta_i\) denote the abilities, \(\delta\) is the tendency of the occurrence of ties and \(\eta\) is the home-advantage effect. The weighting parameter \(w_T\) governs the importance of ties and can be chosen in the argument wgt.ties.

A joint maximum likelihood (JML) estimation is applied for simulataneous estimation of \(\eta\), \(\delta\) and all \(\theta_i\) parameters. In the Rasch model, it was shown that JML can result in biased parameter estimates. The \(\varepsilon\)-adjustment approach has been proposed to reduce the bias in parameter estimates (Bertoli-Bersotti, Lando & Punzo, 2014). This estimation approach is adapted to the Bradley-Terry model in the btm function. To this end, the likelihood function is modified for the purpose of bias reduction. It can be easily shown that there exist sufficient statistics for \(\eta\), \(\delta\) and all \(\theta_i\) parameters. In the \(\varepsilon\)-adjustment approach, the sufficient statistic for the \(\theta_i\) parameter is modified. In JML estimation of the Bradley-Terry model, \(S_i=\sum_{j \ne i} ( x_{ij} + x_{ji} )\) is a sufficient statistic for \(\theta_i\). Let \(M_i\) the maximum score for person \(i\) which is the number of \(x_{ij}\) terms appearing in \(S_i\). In the \(\varepsilon\)-adjustment approach, the sufficient statistic \(S_i\) is modified to $$S_{i, \varepsilon}=\varepsilon + \frac{M_i - 2 \varepsilon}{M_i} S_i $$ and \(S_{i, \varepsilon}\) instead of \(S_{i}\) is used in JML estimation. Hence, original scores \(S_i\) are linearly transformed for all persons \(i\).

Value

List with following entries

pars

Parameter summary for \(\eta\) and \(\delta\)

effects

Parameter estimates for \(\theta\) and outfit and infit statistics

summary.effects

Summary of \(\theta\) parameter estimates

mle.rel

MLE reliability, also known as separation reliability

sepG

Separation index \(G\)

probs

Estimated probabilities

data

Used dataset with integer identifiers

fit_judges

Fit statistics (outfit and infit) for judges if judge is provided. In addition, average agreement of the rating with the mode of the ratings is calculated for each judge (at least three ratings per dyad has to be available for computing the agreement).

residuals

Unstandardized and standardized residuals for each observation

References

Bertoli-Barsotti, L., Lando, T., & Punzo, A. (2014). Estimating a Rasch Model via fuzzy empirical probability functions. In D. Vicari, A. Okada, G. Ragozini & C. Weihs (Eds.). Analysis and Modeling of Complex Data in Behavioral and Social Sciences. Springer. doi: 10.1007/978-3-319-06692-9_4

Hunter, D. R. (2004). MM algorithms for generalized Bradley-Terry models. Annals of Statistics, 32, 384-406. doi: 10.1214/aos/1079120141

See also

Examples