Nonparametric Estimation of Conditional Covariances of Item Pairs

This function estimates conditional covariances of itempairs (Stout, Habing, Douglas & Kim, 1996; Zhang & Stout, 1999a). The function is used for the estimation of the DETECT index.

ccov.np(data, score, bwscale=1.1, thetagrid=seq(-3, 3, len=200),
    progress=TRUE, scale_score=TRUE)

Arguments

data
An \(N \times I\) data frame of dichotomous responses. Missing responses are allowed.

score
An ability estimate, e.g. the WLE

bwscale
Bandwidth factor for calculation of conditional covariance. The bandwidth used in the estimation is bwscale times \(N^{-1/5}\).

thetagrid
A vector which contains theta values where conditional covariances are evaluated.

progress
Display progress?

scale_score
Logical indicating whether score should be z standardized in advance of the calculation of conditional covariances

References

Stout, W., Habing, B., Douglas, J., & Kim, H. R. (1996). Conditional covariance-based nonparametric multidimensionality assessment. Applied Psychological Measurement, 20, 331-354.

Zhang, J., & Stout, W. (1999). Conditional covariance structure of generalized compensatory multidimensional items, Psychometrika, 64, 129-152.

Note

This function is used in conf.detect and expl.detect.