Jackknife Equating Error in Generalized Logistic Rasch Model

This function estimates the linking error in linking based on Jackknife (Monseur & Berezner, 2007).

equating.rasch.jackknife(pars.data, display=TRUE,
   se.linkerror=FALSE, alpha1=0, alpha2=0)

Arguments

pars.data
Data frame with four columns: jackknife unit (1st column), item parameter study 1 (2nd column), item parameter study 2 (3rd column), item (4th column)

display
Display progress?

se.linkerror
Compute standard error of the linking error

alpha1
Fixed \(\alpha_1\) parameter in the generalized item response model

alpha2
Fixed \(\alpha_2\) parameter in the generalized item response model

Value

A list with following entries:

pars.data

Used item parameters

itemunits

Used units for jackknife

descriptives

Descriptives for Jackknife. linkingerror.jackknife is the estimated linking error.

References

Monseur, C., & Berezner, A. (2007). The computation of equating errors in international surveys in education. Journal of Applied Measurement, 8, 323-335.

See also

Examples

# NOT RUN {
#############################################################################
# EXAMPLE 1: Linking errors PISA study
#############################################################################

data(data.pisaPars)
pars <- data.pisaPars

# Linking error: Jackknife unit is the testlet
vars <- c("testlet","study1","study2","item")
res1 <- sirt::equating.rasch.jackknife(pars[, vars])
res1$descriptives
  ##   N.items N.units      shift        SD linkerror.jackknife SE.SD.jackknife
  ## 1      25       8 0.09292838 0.1487387          0.04491197      0.03466309

# Linking error: Jackknife unit is the item
res2 <- sirt::equating.rasch.jackknife(pars[, vars ] )
res2$descriptives
  ##   N.items N.units      shift        SD linkerror.jackknife SE.SD.jackknife
  ## 1      25      25 0.09292838 0.1487387          0.02682839      0.02533327
# }