Scoring Persons and Items in the ISOP Model

This function does the scoring in the isotonic probabilistic model (Scheiblechner, 1995, 2003, 2007). Person parameters are ordinally scaled but the ISOP model also allows specific objective (ordinal) comparisons for persons (Scheiblechner, 1995).

isop.scoring(dat,score.itemcat=NULL)

Arguments

dat
Data frame with dichotomous or polytomous item responses

score.itemcat
Optional data frame with scoring points for every item and every category (see Example 2).

Details

This function extracts the scoring rule of the ISOP model (if score.itemcat !=NULL) and calculates the modified percentile score for every person. The score \(s_{ik}\) for item \(i\) and category \(k\) is calculated as $$ s_{ik}=\sum_{j=0}^{k-1} f_{ij} - \sum_{j=k+1}^K f_{ij}=P( X_i < k ) - P( X_i > k ) $$ where \(f_{ik}\) is the relative frequency of item \(i\) in category \(k\) and \(K\) is the maximum category. The modified percentile score \(\rho_p\) for subject \(p\) (mpsc in person) is defined by $$ \rho_p=\frac{1}{I} \sum_{i=1}^I \sum_{j=0}^K s_{ik} \mathbf{1}( X_{pi}=k ) $$ Note that for dichotomous items, the sum score is a sufficient statistic for \(\rho_p\) but this is not the case for polytomous items. The modified percentile score \(\rho_p\) ranges between -1 and 1.

The modified item P-score \(\rho_i\) (Scheiblechner, 2007, p. 52) is defined by $$ \rho_i=\frac{1}{I-1} \cdot \sum_j \left[ P( X_j < X_i ) - P( X_j > X_i ) \right ] $$

Value

A list with following entries:

person

A data frame with person parameters. The modified percentile score \(\rho_p\) is denoted by mpsc.

item

Item statistics and scoring parameters. The item P-scores \(\rho_i\) are labeled as pscore.

p.itemcat

Frequencies for every item category

score.itemcat

Scoring points for every item category

distr.fct

Empirical distribution function

References

Scheiblechner, H. (1995). Isotonic ordinal probabilistic models (ISOP). Psychometrika, 60, 281-304.

Scheiblechner, H. (2003). Nonparametric IRT: Scoring functions and ordinal parameter estimation of isotonic probabilistic models (ISOP). Technical Report, Philipps-Universitaet Marburg.

Scheiblechner, H. (2007). A unified nonparametric IRT model for d-dimensional psychological test data (d-ISOP). Psychometrika, 72, 43-67.

See also

Examples

#############################################################################
# EXAMPLE 1: Dataset Reading
#############################################################################

data( data.read )
dat <- data.read

# Scoring according to the ISOP model
msc <- sirt::isop.scoring( dat )
# plot student scores
boxplot( msc$person$mpsc ~ msc$person$score )

#############################################################################
# EXAMPLE 2: Dataset students from CDM package | polytomous items
#############################################################################

library("CDM")
data( data.Students, package="CDM")
dat <- stats::na.omit(data.Students[, -c(1:2) ])

# Scoring according to the ISOP model
msc <- sirt::isop.scoring( dat )
# plot student scores
boxplot( msc$person$mpsc ~ msc$person$score )

# scoring with known scoring rule for activity items
items <- paste0( "act", 1:5 )
score.itemcat <- msc$score.itemcat
score.itemcat <- score.itemcat[ items, ]
msc2 <- sirt::isop.scoring( dat[,items], score.itemcat=score.itemcat )