Testing the ISOP Model

This function performs tests of the W1 axiom of the ISOP model (Scheiblechner, 2003). Standard errors of the corresponding \(W1_i\) statistics are obtained by Jackknife.

isop.test(data, jackunits=20, weights=rep(1, nrow(data)))

# S3 method for isop.test
summary(object,...)

Arguments

data
Data frame with item responses

jackunits
A number of Jackknife units (if an integer is provided as the argument value) or a vector in the Jackknife units are already defined.

weights
Optional vector of sampling weights

object
Object of class isop.test

...
Further arguments to be passed

Value

A list with following entries

itemstat

Data frame with test and item statistics for the W1 axiom. The \(W1_i\) statistic is denoted as est while se is the corresponding standard error of the statistic. The sample size per item is N and M denotes the item mean.

Es

Number of concordances per item

Ed

Number of disconcordances per item

The W1_i statistics are printed by the summary method.
References

Scheiblechner, H. (2003). Nonparametric IRT: Testing the bi-isotonicity of isotonic probabilistic models (ISOP). Psychometrika, 68, 79-96.

See also

Examples

#############################################################################
# EXAMPLE 1: ISOP model data.Students
#############################################################################

data(data.Students, package="CDM")
dat <- data.Students[, paste0("act",1:5) ]
dat <- dat[1:300, ]    # select first 300 students

# perform the ISOP test
mod <- sirt::isop.test(dat)
summary(mod)
  ## -> W1i statistics
  ##     parm   N     M   est    se      t
  ##   1 test 300    NA 0.430 0.036 11.869
  ##   2 act1 278 0.601 0.451 0.048  9.384
  ##   3 act2 275 0.473 0.473 0.035 13.571
  ##   4 act3 274 0.277 0.352 0.098  3.596
  ##   5 act4 291 1.320 0.381 0.054  7.103
  ##   6 act5 276 0.460 0.475 0.042 11.184