Multidimensional IRT Copula Model

This function handles local dependence by specifying copulas for residuals in multidimensional item response models for dichotomous item responses (Braeken, 2011; Braeken, Tuerlinckx & de Boeck, 2007; Schroeders, Robitzsch & Schipolowski, 2014). Estimation is allowed for item difficulties, item slopes and a generalized logistic link function (Stukel, 1988).

The function rasch.copula3 allows the estimation of multidimensional models while rasch.copula2 only handles unidimensional models.

rasch.copula2(dat, itemcluster, copula.type="bound.mixt",
    progress=TRUE, mmliter=1000, delta=NULL,
    theta.k=seq(-4, 4, len=21), alpha1=0, alpha2=0,
    numdiff.parm=1e-06,  est.b=seq(1, ncol(dat)),
    est.a=rep(1, ncol(dat)), est.delta=NULL, b.init=NULL, a.init=NULL,
    est.alpha=FALSE, glob.conv=0.0001, alpha.conv=1e-04, conv1=0.001,
    dev.crit=.2, increment.factor=1.01)

rasch.copula3(dat, itemcluster, dims=NULL, copula.type="bound.mixt",
    progress=TRUE, mmliter=1000, delta=NULL,
    theta.k=seq(-4, 4, len=21), alpha1=0, alpha2=0,
    numdiff.parm=1e-06,  est.b=seq(1, ncol(dat)),
    est.a=rep(1, ncol(dat)), est.delta=NULL, b.init=NULL, a.init=NULL,
    est.alpha=FALSE, glob.conv=0.0001, alpha.conv=1e-04, conv1=0.001,
    dev.crit=.2, rho.init=.5, increment.factor=1.01)

# S3 method for rasch.copula2
summary(object, file=NULL, digits=3, ...)
# S3 method for rasch.copula3
summary(object, file=NULL, digits=3, ...)

# S3 method for rasch.copula2
anova(object,...)
# S3 method for rasch.copula3
anova(object,...)

# S3 method for rasch.copula2
logLik(object,...)
# S3 method for rasch.copula3
logLik(object,...)

# S3 method for rasch.copula2
IRT.likelihood(object,...)
# S3 method for rasch.copula3
IRT.likelihood(object,...)

# S3 method for rasch.copula2
IRT.posterior(object,...)
# S3 method for rasch.copula3
IRT.posterior(object,...)

Arguments

dat
An \(N \times I\) data frame. Cases with only missing responses are removed from the analysis.

itemcluster
An integer vector of length \(I\) (number of items). Items with the same integers define a joint item cluster of (positively) locally dependent items. Values of zero indicate that the corresponding item is not included in any item cluster of dependent responses.

dims
A vector indicating to which dimension an item is allocated. The default is that all items load on the first dimension.

copula.type
A character or a vector containing one of the following copula types: bound.mixt (boundary mixture copula), cook.johnson (Cook-Johnson copula) or frank (Frank copula) (see Braeken, 2011). The vector copula.type must match the number of different itemclusters. For every itemcluster, a different copula type may be specified (see Examples).

progress
Print progress? Default is TRUE.

mmliter
Maximum number of iterations.

delta
An optional vector of starting values for the dependency parameter delta.

theta.k
Discretized trait distribution

alpha1
alpha1 parameter in the generalized logistic item response model (Stukel, 1988). The default is 0 which leads together with alpha2=0 to the logistic link function.

alpha2
alpha2 parameter in the generalized logistic item response model

numdiff.parm
Parameter for numerical differentiation

est.b
Integer vector of item difficulties to be estimated

est.a
Integer vector of item discriminations to be estimated

est.delta
Integer vector of length length(itemcluster). Nonzero integers correspond to delta parameters which are estimated. Equal integers indicate parameter equality constraints.

b.init
Initial \(b\) parameters

a.init
Initial \(a\) parameters

est.alpha
Should both alpha parameters be estimated? Default is FALSE.

glob.conv
Convergence criterion for all parameters

alpha.conv
Maximal change in alpha parameters for convergence

conv1
Maximal change in item parameters for convergence

dev.crit
Maximal change in the deviance. Default is .2.

rho.init
Initial value for off-diagonal elements in correlation matrix

increment.factor
A numeric value larger than one which controls the size of increments in iterations. To stabilize convergence, choose values 1.05 or 1.1 in some situations.

object
Object of class rasch.copula2 or rasch.copula3

file
Optional file name for summary output

digits
Number of digits after decimal in summary output

...
Further arguments to be passed

Value

A list with following entries

N.itemclusters

Number of item clusters

item

Estimated item parameters

iter

Number of iterations

dev

Deviance

delta

Estimated dependency parameters \(\delta\)

b

Estimated item difficulties

a

Estimated item slopes

mu

Mean

sigma

Standard deviation

alpha1

Parameter \(\alpha_1\) in the generalized item response model

alpha2

Parameter \(\alpha_2\) in the generalized item response model

ic

Information criteria

theta.k

Discretized ability distribution

pi.k

Fixed \(\theta\) distribution

deviance

Deviance

pattern

Item response patterns with frequencies and posterior distribution

person

Data frame with person parameters

datalist

List of generated data frames during estimation

EAP.rel

Reliability of the EAP

copula.type

Type of copula

summary.delta

Summary for estimated \(\delta\) parameters

f.qk.yi

Individual posterior

f.yi.qk

Individual likelihood

...

Further values

References

Braeken, J. (2011). A boundary mixture approach to violations of conditional independence. Psychometrika, 76(1), 57-76. doi: 10.1007/s11336-010-9190-4

Braeken, J., Kuppens, P., De Boeck, P., & Tuerlinckx, F. (2013). Contextualized personality questionnaires: A case for copulas in structural equation models for categorical data. Multivariate Behavioral Research, 48(6), 845-870. doi: 10.1080/00273171.2013.827965

Braeken, J., & Tuerlinckx, F. (2009). Investigating latent constructs with item response models: A MATLAB IRTm toolbox. Behavior Research Methods, 41(4), 1127-1137.

Braeken, J., Tuerlinckx, F., & De Boeck, P. (2007). Copula functions for residual dependency. Psychometrika, 72(3), 393-411. doi: 10.1007/s11336-007-9005-4

Schroeders, U., Robitzsch, A., & Schipolowski, S. (2014). A comparison of different psychometric approaches to modeling testlet structures: An example with C-tests. Journal of Educational Measurement, 51(4), 400-418. doi: 10.1111/jedm.12054

Stukel, T. A. (1988). Generalized logistic models. Journal of the American Statistical Association, 83(402), 426-431. doi: 10.1080/01621459.1988.10478613

See also

Examples