Estimation of the Partial Credit Model using the Eigenvector Method

This function performs the eigenvector approach to estimate item parameters which is based on a pairwise estimation approach (Garner & Engelhard, 2002). No assumption about person parameters is required for item parameter estimation. Statistical inference is performed by Jackknifing. If a group identifier is provided, tests for differential item functioning are performed.

rasch.evm.pcm(dat, jackunits=20, weights=NULL, pid=NULL,
    group=NULL, powB=2, adj_eps=0.3, progress=TRUE )

# S3 method for rasch.evm.pcm
summary(object, digits=3, file=NULL, ...)

# S3 method for rasch.evm.pcm
coef(object,...)

# S3 method for rasch.evm.pcm
vcov(object,...)

Arguments

dat
Data frame with dichotomous or polytomous item responses

jackunits
A number of Jackknife units (if an integer is provided as the argument value) or a vector in which the Jackknife units are already defined.

weights
Optional vector of sample weights

pid
Optional vector of person identifiers

group
Optional vector of group identifiers. In this case, item parameters are group wise estimated and tests for differential item functioning are performed.

powB
Power created in \(B\) matrix which is the basis of parameter estimation

adj_eps
Adjustment parameter for person parameter estimation (see mle.pcm.group)

progress
An optional logical indicating whether progress should be displayed

object
Object of class rasch.evm.pcm

digits
Number of digits after decimals for rounding in summary.

file
Optional file name if summary should be sunk into a file.

...
Further arguments to be passed

Value

A list with following entries

item

Data frame with item parameters. The item parameter estimate is denoted by est while a Jackknife bias-corrected estimate is est_jack. The Jackknife standard error is se.

b

Item threshold parameters

person

Data frame with person parameters obtained (MLE)

B

Paired comparison matrix

D

Transformed paired comparison matrix

coef

Vector of estimated coefficients

vcov

Covariance matrix of estimated item parameters

JJ

Number of jackknife units

JJadj

Reduced number of jackknife units

powB

Used power of comparison matrix \(B\)

maxK

Maximum number of categories per item

G

Number of groups

desc

Some descriptives

difstats

Statistics for differential item functioning if group is provided as an argument

References

Choppin, B. (1985). A fully conditional estimation procedure for Rasch Model parameters. Evaluation in Education, 9, 29-42.

Garner, M., & Engelhard, G. J. (2002). An eigenvector method for estimating item parameters of the dichotomous and polytomous Rasch models. Journal of Applied Measurement, 3, 107-128.

Wang, J., & Engelhard, G. (2014). A pairwise algorithm in R for rater-mediated assessments. Rasch Measurement Transactions, 28(1), 1457-1459.

See also

Examples