PROX Estimation Method for the Rasch Model

This function estimates the Rasch model using the PROX algorithm (cited in Wright & Stone, 1999).

rasch.prox(dat, dat.resp=1 - is.na(dat), freq=rep(1,nrow(dat)),
    conv=0.001, maxiter=30, progress=FALSE)

Arguments

dat
An \(N \times I\) data frame of dichotomous response data. NAs are not allowed and must be indicated by zero entries in the response indicator matrix dat.resp.

dat.resp
An \(N \times I\) indicator data frame of nonmissing item responses.

freq
A vector of frequencies (or weights) of all rows in data frame dat.

conv
Convergence criterion for item parameters

maxiter
Maximum number of iterations

progress
Display progress?

Value

A list with following entries

b

Estimated item difficulties

theta

Estimated person abilities

iter

Number of iterations

sigma.i

Item standard deviations

sigma.n

Person standard deviations

References

Wright, B., & Stone, W. (1999). Measurement Essentials. Wilmington: Wide Range.

Examples

#############################################################################
# EXAMPLE 1: PROX data.read
#############################################################################

data(data.read)
mod <- sirt::rasch.prox( data.read )
mod$b       # item difficulties