Weighted Likelihood Estimation of Person Abilities

This function computes weighted likelihood estimates for dichotomous responses based on the Rasch model (Warm, 1989).

wle.rasch(dat, dat.resp=NULL, b, itemweights=1 + 0 * b,
    theta=rep(0, nrow(dat)), conv=0.001, maxit=200,
    wle.adj=0, progress=FALSE)

Arguments

dat
An \(N \times I\) data frame of dichotomous item responses

dat.resp
Optional data frame with dichotomous response indicators

b
Vector of length \(I\) with fixed item difficulties

itemweights
Optional vector of fixed item discriminations

theta
Optional vector of initial person parameter estimates

conv
Convergence criterion

maxit
Maximal number of iterations

wle.adj
Constant for WLE adjustment

progress
Display progress?

Value

A list with following entries

theta

Estimated weighted likelihood estimate

dat.resp

Data frame with dichotomous response indicators. A one indicates an observed response, a zero a missing response. See also dat.resp in the list of arguments of this function.

p.ia

Matrix with expected item response, i.e. the probabilities \(P(X_{pi}=1|\theta_p )=invlogit( \theta_p - b_i )\).

wle

WLE reliability (Adams, 2005)

References

Adams, R. J. (2005). Reliability as a measurement design effect. Studies in Educational Evaluation, 31, 162-172.

Warm, T. A. (1989). Weighted likelihood estimation of ability in item response theory. Psychometrika, 54, 427-450.

See also

Examples

#############################################################################
# EXAMPLE 1: Dataset Reading
#############################################################################
data(data.read)

# estimate the Rasch model
mod <- sirt::rasch.mml2(data.read)
mod$item

# estmate WLEs
mod.wle <- sirt::wle.rasch( dat=data.read, b=mod$item$b )